Create a first version of WebAPI classes: mockups, interfaces, controllers. Update models and resources. Update view cache. Declare ref between mocks and interfaces.

+++ src/TheInternecineStrife/TheInternecineStrife.ServerSide.Model/Economic/Resource.cs
@@ -43,7 +43,7 @@ namespace TheInternecineStrife.ServerSide.Model.Economic
 
         public float CurrentProfit(float instrument)
         {
-            return (float)(Math.Min(Profit, instrument) * Available / Amount);
+            return Math.Min(Profit, instrument) * Available / Amount;
         }
 
         public float Extract(int workers, float instrument)
